pyproject.toml
[tool.poetry]
name = "django-gov-notify"
version = "0.1.0"
description = ""
authors = ["Nick Smith <nick.smith@torchbox.com>"]

[tool.poetry.dependencies]
python = ">=3.6.1,<3.9"
django = "2.2.12"
notifications-python-client = "5.4.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
flake8 = "^3.8.2"
pre-commit = "^2.4.0"
detect-secrets = "0.13.0"
factory-boy = "^2.12.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
